Splitting commands into segregated, testable classes
