+++ src/Example/Cli/Greet.php
@@ -1,29 +1,16 @@
-#!/usr/bin/env php
 <?php
 
-use Symfony\Component\Console\Application;
+declare(strict_types=1);
+
+namespace Example\Cli;
+
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
-require_once __DIR__ . '/../vendor/autoload.php';
-
-$app = new Application('Example application - inlined version');
-
-$app->add(new class extends Command {
-    protected function configure()
-    {
-        $this->setName('example:say-hello');
-    }
-
-    protected function execute(InputInterface $input, OutputInterface $output)
-    {
-        $output->writeln('<info>Hello World!</info>');
-    }
-});
-
-$app->add(new class extends Command {
+final class Greet extends Command
+{
     protected function configure()
     {
         $this->setName('example:greet');
@@ -37,6 +24,4 @@ $app->add(new class extends Command {
             $input->getArgument('name')
         ));
     }
-});
-
-$app->run();
+}
